refactored mapbox parsing into seperate class and make house number optional
