+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Minutes in Dates are displayed correctly 
 - Display date after 24 Hours in Alarmcounter
+- Make house number in Mapboxgeocoding result optional
 
 ## [1.0.1](https://github.com/denniskniep/alarmcontrol/compare/v1.0.0-24...v1.0.1-30) - 2019-07-07
 ### Added
