+++ src/main/java/com/alarmcontrol/server/maps/mapbox/geocoding/MapboxGeocodingService.java
@@ -5,12 +5,8 @@ import com.alarmcontrol.server.maps.CachingRestService.Request;
 import com.alarmcontrol.server.maps.CachingRestService.Response;
 import com.alarmcontrol.server.maps.GeocodingResult;
 import com.alarmcontrol.server.maps.GeocodingService;
-import com.jayway.jsonpath.JsonPath;
 import java.net.URI;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpEntity;
@@ -18,7 +14,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
 import org.springframework.web.util.UriComponentsBuilder;
 
 @Service
@@ -66,33 +61,6 @@ public class MapboxGeocodingService implements GeocodingService {
   }
 
   private GeocodingResult parse(String json){
-    String lon = JsonPath.<Double>read(json, "$.features[0].center[0]").toString();
-    String lat = JsonPath.<Double>read(json, "$.features[0].center[1]").toString();
-
-    String houseNumber = JsonPath.read(json, "$.features[0].address");
-    String road = JsonPath.read(json, "$.features[0].text");
-
-    List<String> context =  JsonPath.read(json, "$.features[0].context[*].text");
-
-    String cityDistrict = "";
-    String city = "";
-
-    //directly in the main town
-    if(context.size() == 4){
-      city = context.get(1);
-    }else if(context.size() > 4){
-      city = context.get(2);
-      cityDistrict = context.get(0);
-    }
-
-    String addressInfo1 = join(" ", road, houseNumber);
-    String addressInfo2 = join("-", city, cityDistrict);
-
-    return new GeocodingResult(json, lat, lon, addressInfo1, addressInfo2);
-  }
-
-  private String join(String delimiter, String ...elements) {
-    List<String> filteredElements = Arrays.stream(elements).filter(e -> !StringUtils.isEmpty(e)).collect(Collectors.toList());
-    return String.join(delimiter, filteredElements);
+   return new MapboxParser().parse(json);
   }
 }
